Rename layout_tab to inspector_screen_details_tab
