+++ packages/devtools_app/lib/src/inspector/flutter/inspector_screen_details_tab.dart
